Conflicts:
	besto_prj/target/m2e-wtp/web-resources/META-INF/MANIFEST.MF
	besto_prj/target/m2e-wtp/web-resources/META-INF/maven/com.kosmo/member/pom.properties

+++ besto_prj/src/main/java/com/com/game/RecordRateVO.java
@@ -2,59 +2,59 @@ package com.com.game;
 
 public class RecordRateVO {
 	private int matchSeq;
-	private int recoderate00;
-	private int recoderate10;
-	private int recoderate20;
-	private int recoderate11;
-	private int recoderate22;
-	private int recoderate01;
-	private int recoderate02;
+	private float recoderate00;
+	private float recoderate10;
+	private float recoderate20;
+	private float recoderate11;
+	private float recoderate22;
+	private float recoderate01;
+	private float recoderate02;
 	public int getMatchSeq() {
 		return matchSeq;
 	}
 	public void setMatchSeq(int matchSeq) {
 		this.matchSeq = matchSeq;
 	}
-	public int getRecoderate00() {
+	public float getRecoderate00() {
 		return recoderate00;
 	}
-	public void setRecoderate00(int recoderate00) {
+	public void setRecoderate00(float recoderate00) {
 		this.recoderate00 = recoderate00;
 	}
-	public int getRecoderate10() {
+	public float getRecoderate10() {
 		return recoderate10;
 	}
-	public void setRecoderate10(int recoderate10) {
+	public void setRecoderate10(float recoderate10) {
 		this.recoderate10 = recoderate10;
 	}
-	public int getRecoderate20() {
+	public float getRecoderate20() {
 		return recoderate20;
 	}
-	public void setRecoderate20(int recoderate20) {
+	public void setRecoderate20(float recoderate20) {
 		this.recoderate20 = recoderate20;
 	}
-	public int getRecoderate11() {
+	public float getRecoderate11() {
 		return recoderate11;
 	}
-	public void setRecoderate11(int recoderate11) {
+	public void setRecoderate11(float recoderate11) {
 		this.recoderate11 = recoderate11;
 	}
-	public int getRecoderate22() {
+	public float getRecoderate22() {
 		return recoderate22;
 	}
-	public void setRecoderate22(int recoderate22) {
+	public void setRecoderate22(float recoderate22) {
 		this.recoderate22 = recoderate22;
 	}
-	public int getRecoderate01() {
+	public float getRecoderate01() {
 		return recoderate01;
 	}
-	public void setRecoderate01(int recoderate01) {
+	public void setRecoderate01(float recoderate01) {
 		this.recoderate01 = recoderate01;
 	}
-	public int getRecoderate02() {
+	public float getRecoderate02() {
 		return recoderate02;
 	}
-	public void setRecoderate02(int recoderate02) {
+	public void setRecoderate02(float recoderate02) {
 		this.recoderate02 = recoderate02;
 	}
